pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["Abílio Azevedo <abilio10@gmail.com>"]
readme = "README.md"
packages = [{include = "todo_list"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
